Finish a working computerPlay function
With this commit, variables are declared for both the playerSelection
and the computerSelection, but neither is set.
The computerPlay function is completed, so that based on
which number is selected at random, it will set the computerSelection
variable to either paper, scissors, or rock.
This is tested via console.log and works as expected.
However, I can now see that by refactoring the way this function works,
I can ultimately do away with the two initial variables.