Refactor computerPlay function
By the last commit, I could see that the computerPlay function
had redundancies in that it didn't need a global variable to function.
With this commit, the code is refactored, so that the switch statement
returns values for rock paper scissors, rather than setting them
to a global variable.
The player and computer selection variables are also removed.
This is tested to be working via console.log.
The names for rock paper and scissors are also changed
to give freshness to the game and are capitalized for style.