Write first part of playRound function
To begin with, the playRound needs to standardize user input.
To do this, the playerSelection variable is hit
with the toLowerCase function.
In adding this function, it becomes apparent that computerSelection
should start as lowercase also then.
So with this commit, return values from case statement
in computerPlay function are changed to lowercase.
It also becomes apparent that the default "Error" catch
should have a place in this next function and so the first line
of the playRound function handles this possibility.