Finish Game function, console test and decorate
With this commit, the game function is tested and working.
It could definitely be reworked to use numbers instead of words
to calculate score and with some sort of function based counter,
rather than using the String.IndexOf() trick that I used instead.
However, with the way it sits, the game works
and each response is unique to the play,
which actually gives the game more of a feel.
Before adding HTML or CSS to this though, I will definitely look over it
and clean it up, removing any redundancies
and replacing any dangers/mess.