Add Restart Button and associated functionality
With this commit, once a game hits completion, a function is called,
which displays a button to click to play the game again
from the start.
There is also now a top paragraph with instructions on how to play,
which will disappear with the appearance of the button.
Once the game is restarted, via this Play Again button, the scores both
reset to 0 and the results and score paragraphs become empty.
The top paragraph also returns to visibility.